Complete contract payment

+++ resources/views/contracts/payments/index.blade.php
@@ -168,7 +168,11 @@
                             {{ $loop->index + 1 }}
                         </td>
                         <td class="table-tr-td border-t-0 border-x-0">
-                            {{ jdate($payment->date)->format('Y/m/d') }}
+                            @if(!is_null($payment->date))
+                                {{ jdate($payment->date)->format('Y/m/d') }}
+                            @else
+                                -
+                            @endif
                         </td>
                         <td class="table-tr-td border-t-0 border-x-0">
                             {{ $payment->text }}
@@ -233,15 +237,17 @@
                                     ویرایش
                                 </a>
 
-                                <button type="button" onclick="destroyPayment({{ $payment->id }})"
-                                        class="table-delete-btn">
-                                    <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
-                                         stroke-width="1.5" stroke="currentColor" class="w-4 h-4 ml-1">
-                                        <path stroke-linecap="round" stroke-linejoin="round"
-                                              d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"/>
-                                    </svg>
-                                    حذف
-                                </button>
+                                @if(!$payment->confirm)
+                                    <button type="button" onclick="destroyPayment({{ $payment->id }})"
+                                            class="table-delete-btn">
+                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
+                                             stroke-width="1.5" stroke="currentColor" class="w-4 h-4 ml-1">
+                                            <path stroke-linecap="round" stroke-linejoin="round"
+                                                  d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"/>
+                                        </svg>
+                                        حذف
+                                    </button>
+                                @endif
                             </div>
                         </td>
                     </tr>
@@ -266,6 +272,7 @@
         $contractPrice = 0;
         $paymentPrice = 0;
         $leftPrice = 0;
+        $collectionPrice = 0;
 
         foreach ($contract->products as $product) {
             $contractPrice += $product->price * $product->quantity;
@@ -277,6 +284,10 @@
             } else {
                 $paymentPrice += $payment2->price;
             }
+
+            if (is_null($payment2->account_id) && $payment2->cash_type == 'check') {
+                $collectionPrice += $payment2->price;
+            }
         }
 
         $tax = $contractPrice * 9 / 100;
@@ -286,7 +297,7 @@
         $leftTaxPrice = $contractTaxPrice - $paymentPrice;
     @endphp
 
-    <div class="mt-8 grid grid-cols-3 gap-4">
+    <div class="mt-8 grid grid-cols-4 gap-4">
         <div class="p-4 rounded-lg shadow bg-indigo-500">
             @if($contract->type == 'official')
                 <p class="text-base text-white text-center font-bold">
@@ -298,6 +309,7 @@
                 </p>
             @endif
         </div>
+
         <div class="p-4 rounded-lg shadow bg-green-500">
             <p class="text-base text-white text-center font-bold">
                 مجموع پرداخت‌ها : {{ number_format($paymentPrice) }} تومان
@@ -316,10 +328,19 @@
             @endif
         </div>
 
+        <div class="p-4 rounded-lg shadow bg-yellow-500">
+            <p class="text-base text-white text-center font-bold">
+                وصول نشده‌ها : {{ number_format($collectionPrice) }} تومان
+            </p>
+        </div>
+
         @if($contract->type == 'official')
-            <div>
+            <div class="space-y-4">
                 <p class="text-center text-sm underline underline-offset-4 text-gray-600">
                     مبلغ کل قرارداد بدون ارزش افزوده : {{ number_format($contractPrice) }} تومان
+                </p>
+                <p class="text-center text-sm underline underline-offset-4 text-gray-600">
+                    مبلغ ارزش افزوده : {{ number_format($tax) }} تومان
                 </p>
             </div>
         @endif
